refactor(models): typed Flags on SponsorshipSet/SponsorshipTransfer
Shadow the base uint Flags with SponsorshipSetFlags?/SponsorshipTransferFlags?
following the established Batch/Payment pattern; wire format unchanged.

+++ Tests/Xrpl.Tests/Integration/transactions/TestISponsorship.cs
@@ -148,7 +148,7 @@ public class TestISponsorship
         {
             Account = sponsor.ClassicAddress,
             Sponsee = sponsee.ClassicAddress,
-            Flags = (uint)SponsorshipSetFlags.tfDeleteObject,
+            Flags = SponsorshipSetFlags.tfDeleteObject,
         };
         delete = await client.Autofill(delete);
         ValidateResult(await client.SubmitAndWait(delete, sponsor, true));
